requirements.txt
antlr4-python3-runtime==4.8
bitarray==2.7.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
Cython==0.29.34
Deprecated==1.2.13
fairseq==0.12.2
filelock==3.12.0
fugashi==1.2.1
hydra-core==1.0.7
idna==3.4
jaconv==0.3.4
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
mecab==0.996.3
mozcpy==0.1.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
omegaconf==2.0.6
OpenCC==1.1.6
pandas==2.0.1
plac==1.3.5
portalocker==2.7.0
pycparser==2.21
pykakasi==2.2.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
sacrebleu==2.3.1
six==1.16.0
sympy==1.11.1
tabulate==0.9.0
torch==2.0.1
torchaudio==2.0.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
unidic==1.1.0
urllib3==2.0.2
wanakana-python==1.2.2
wasabi==0.10.1
wrapt==1.15.0
